rename suggestions display
